fix: add more checks for correct ANDROID_HOME
- if ANDROID_HOME is set, check if the path is valid and clean it up if the path is not existing
- add more path checks when updating Android SDK tools
- add better coloring of messages

+++ native-script.ps1
@@ -43,7 +43,7 @@ function Install($programName, $message, $script, $shouldExit) {
 		} until ($script:answer -eq "y" -or $script:answer -eq "n" -or $script:answer -eq "a")
 
 		if ($script:answer -eq "n") {
-			Write-Host -ForegroundColor Yellow "You have chosen not to install $($programName). Some features of NativeScript may not work correctly if you haven't already installed it"
+			Write-Host -ForegroundColor Yellow "WARNING: You have chosen not to install $($programName). Some features of NativeScript may not work correctly if you haven't already installed it"
 			return
 		}
 	}
@@ -76,12 +76,18 @@ Install "Google Chrome" "Installing Google Chrome (required to debug NativeScrip
 
 Install "Java Development Kit" "Installing Java Development Kit" "choco upgrade jdk8 --force"
 
+$androidHomePathExists = Test-Path $env:ANDROID_HOME
+if($androidHomePathExists -eq $False){
+	[Environment]::SetEnvironmentVariable("ANDROID_HOME",$null,"User")
+}
+
 Install "Android SDK" "Installing Android SDK" "cinst android-sdk --force --yes"
 
 refreshenv
 # setup environment
 
 if (!$env:ANDROID_HOME) {
+	Write-Host -ForegroundColor DarkYellow "Setting up ANDROID_HOME"
 	# in case the user has `android` in the PATH, use it as base for setting ANDROID_HOME
 	$androidExecutableEnvironmentPath = Get-Command android -ErrorAction SilentlyContinue | Select-Object -ExpandProperty Definition
 	if ($androidExecutableEnvironmentPath -ne $null) {
@@ -105,19 +111,28 @@ if (!$env:JAVA_HOME) {
 	refreshenv
 }
 
+Write-Host -ForegroundColor DarkYellow "Setting up Android SDK..."
+
 # Update android-sdk tools in order to have avdmanager available and create AVDs later
 # NOTE: This step can be removed once chocolatey provides later version of Android SDK
 $androidToolsPath = [io.path]::combine($env:ANDROID_HOME, "tools")
 $androidToolsOldPath = [io.path]::combine($env:ANDROID_HOME, "toolsOld")
-Copy-Item "$androidToolsPath" "$androidToolsOldPath" -recurse
-echo y | cmd /c "%ANDROID_HOME%\toolsOld\bin\sdkmanager.bat" "tools"
-Remove-Item "$androidToolsOldPath" -Force -Recurse
+
+$androidToolsPathExists = Test-Path $androidToolsPath
+if($androidToolsPathExists -eq $True){
+	Write-Host -ForegroundColor DarkYellow "Updating Android SDK tools..."
+	Copy-Item "$androidToolsPath" "$androidToolsOldPath" -recurse
+	echo y | cmd /c "%ANDROID_HOME%\toolsOld\bin\sdkmanager.bat" "tools"
+	Remove-Item "$androidToolsOldPath" -Force -Recurse
+} else {
+	Write-Host -ForegroundColor Red "ERROR: Failed to update Android SDK tools. This is a blocker to install default emulator, so please update manually once this installation has finished."
+}
 
 # add repositories.cfg if it is not created
 $repositoriesConfigPath = [io.path]::combine($env:USERPROFILE, ".android", "repositories.cfg")
 $pathExists = Test-Path $repositoriesConfigPath
 if($pathExists -eq $False){
-	Write-Host "Creating file $repositoriesConfigPath ..."
+	Write-Host -ForegroundColor DarkYellow "Creating file $repositoriesConfigPath ..."
 	New-Item $repositoriesConfigPath -type file
 }
 
@@ -125,18 +140,17 @@ if($pathExists -eq $False){
 # following commands are separated in case of having to answer to license agreements
 $androidExecutable = [io.path]::combine($env:ANDROID_HOME, "tools", "bin", "sdkmanager")
 
-Write-Host "Setting up Android SDK..."
-Write-Host "Setting up Android SDK platform-tools..."
+Write-Host -ForegroundColor DarkYellow "Setting up Android SDK platform-tools..."
 echo y | cmd /c "$androidExecutable" "platform-tools"
-Write-Host "Setting up Android SDK build-tools;25.0.2..."
+Write-Host -ForegroundColor DarkYellow "Setting up Android SDK build-tools;25.0.2..."
 echo y | cmd /c "$androidExecutable" "build-tools;25.0.2"
-Write-Host "Setting up Android SDK platforms;android-25..."
+Write-Host -ForegroundColor DarkYellow "Setting up Android SDK platforms;android-25..."
 echo y | cmd /c "$androidExecutable" "platforms;android-25"
-Write-Host "Setting up Android SDK extras;android;m2repository..."
+Write-Host -ForegroundColor DarkYellow "Setting up Android SDK extras;android;m2repository..."
 echo y | cmd /c "$androidExecutable" "extras;android;m2repository"
-Write-Host "Setting up Android SDK extras;google;m2repository..."
+Write-Host -ForegroundColor DarkYellow "Setting up Android SDK extras;google;m2repository..."
 echo y | cmd /c "$androidExecutable" "extras;google;m2repository"
-Write-Host "FINISHED setting up Android SDK."
+Write-Host -ForegroundColor DarkYellow "FINISHED setting up Android SDK."
 
 # Setup Default Emulator
 iex ((new-object net.webclient).DownloadString($scriptCommonUrl))
